@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 
 // styles
 import 'bootstrap/dist/css/bootstrap.css';
@@ -8,5 +8,8 @@ import './styles/globals.css';
 // main page
 import {IndexPage} from './pages';
 
+const container = document.getElementById('root');
+if (!container) throw new Error("Failed to find root element!");
 
-ReactDOM.render(<IndexPage/>, document.getElementById('root'));
+const root = createRoot(container);
+root.render(<IndexPage />);
