@@ -1,6 +1,6 @@
 import React from "react";
 
-import {Redirect, Route, Switch} from "react-router-dom";
+import {Navigate, Route, Routes} from "react-router-dom";
 
 import {useRouter} from "../../../../lib/hooks/router";
 import UserGroupsPage from "./groups";
@@ -12,27 +12,17 @@ import UserPoliciesPage from "./policies";
 const UserRedirect = ({ subPath }) => {
     const router = useRouter();
     const {userId} = router.params;
-    return <Redirect to={`/auth/users/${userId}${subPath}`}/>;
+    return <Navigate to={`/auth/users/${userId}${subPath}`}/>;
 }
 
 export default function UserPage() {
     return (
-        <Switch>
-            <Route exact path="/auth/users/:userId">
-                <UserRedirect subPath="/groups"/>
-            </Route>
-            <Route exact path="/auth/users/:userId/groups">
-                <UserGroupsPage />
-            </Route>
-            <Route exact path="/auth/users/:userId/policies">
-                <UserPoliciesPage />
-            </Route>
-            <Route exact path="/auth/users/:userId/policies/effective">
-                <UserEffectivePoliciesPage />
-            </Route>
-            <Route exact path="/auth/users/:userId/credentials">
-                <UserCredentialsPage />
-            </Route>
-        </Switch>
+        <Routes>
+            <Route path="" element={<UserRedirect subPath="/groups"/>} />
+            <Route path="groups" element={<UserGroupsPage />} />
+            <Route exact path="policies" element={<UserPoliciesPage />} />
+            <Route exact path="policies/effective" element={<UserEffectivePoliciesPage />} />
+            <Route exact path="credentials" element={<UserCredentialsPage />} />
+        </Routes>
     );
 }
